% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_references.R
\name{mosaic_calc_link}
\alias{mosaic_calc_link}
\title{Services for MOSAIC Calculus book}
\usage{
mosaic_calc_link(chap, label)
}
\arguments{
\item{chap}{Chapter number}

\item{label}{(optional) name of a section or other cross-reference}
}
\description{
Provides links to MOSAIC Calculus links
}
